# .futurxlab/agent-platform-api-spec.yaml
openapi: 3.0.0
info:
  title: Agent Content Generation Platform API
  version: 1.0.0
  description: Internal API for AI-powered content generation using Dify agents
  contact:
    name: FuturX Development Team
    email: dev@futurx.com
servers:
  - url: http://localhost:7860/api
    description: Development server
  - url: https://agent-platform.internal/api
    description: Production server
...
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Authenticate user with password
      description: Validates password and creates session
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - password
              properties:
                password:
                  type: string
                  format: password
                  description: System access password
                  example: "secure_password_123"
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: Session token
                    example: "eyJhbGciOiJIUzI1NiIs..."
                  sessionId:
                    type: string
                    description: Session identifier
                    example: "sess_abc123"
                  expiresAt:
                    type: string
                    format: date-time
                    description: Session expiration time
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout current user
      description: Invalidates current session
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Logged out successfully"
  /auth/session:
    get:
      tags:
        - Authentication
      summary: Check session status
      operationId: checkSession
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Session is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  expiresAt:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # Agent Management Endpoints (#REF-AGENT-01)
  /agents:
    get:
      tags:
        - Agents
      summary: List all configured agents
      operationId: listAgents
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                type: object
                properties:
                  agents:
                    type: array
                    items:
                      $ref: '#/components/schemas/Agent'

    post:
      tags:
        - Agents
      summary: Add new agent configuration
      operationId: createAgent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentCreate'
      responses:
        '201':
          description: Agent created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequestError'
  /agents/validate:
    post:
      tags:
        - Agents
      summary: Validate agent API token
      operationId: validateAgent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - url
                - apiToken
              properties:
                url:
                  type: string
                  format: uri
                apiToken:
                  type: string
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  message:
                    type: string
  /agents/{agentId}:
    get:
      tags:
        - Agents
      summary: Get agent details
      operationId: getAgent
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/agentId'
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags:
        - Agents
      summary: Update agent configuration
      operationId: updateAgent
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/agentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentUpdate'
      responses:
        '200':
          description: Agent updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'

    delete:
      tags:
        - Agents
      summary: Delete agent configuration
      operationId: deleteAgent
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/agentId'
      responses:
        '204':
          description: Agent deleted successfully

  # Task Management Endpoints (#REF-TASK-01)
...
  /generation/start:
    post:
      tags:
        - Generation
      summary: Start content generation
      operationId: startGeneration
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - agentId
                - taskId
                - input
              properties:
                agentId:
                  type: string
                  description: Selected agent ID
                taskId:
                  type: string
                  description: Selected task ID
                input:
                  type: object
                  properties:
                    text:
                      type: string
                      description: User input text
                    fileId:
                      type: string
                      description: Uploaded file ID
      responses:
        '202':
          description: Generation started
          content:
            application/json:
              schema:
                type: object
                properties:
                  generationId:
                    type: string
                  status:
                    type: string
                    enum: [starting, processing]
                  streamUrl:
                    type: string
                    description: URL for SSE stream
  /generation/stream:
    get:
      tags:
        - Generation
      summary: Stream generation results
      operationId: streamGeneration
      security:
        - bearerAuth: []
      parameters:
        - name: generationId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: SSE stream of generation chunks
          content:
            text/event-stream:
              schema:
                type: object
                properties:
                  event:
                    type: string
                    enum: [chunk, complete, error]
                  data:
                    type: string
  /generation/download/{generationId}:
    get:
      tags:
        - Generation
      summary: Download generation result
      operationId: downloadGeneration
      security:
        - bearerAuth: []
      parameters:
        - name: generationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Markdown file
          content:
            text/markdown:
              schema:
                type: string
        '404':
          $ref: '#/components/responses/NotFoundError'

  # File Upload Endpoint (#REF-UPLOAD-01)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    agentId:
      name: agentId
      in: path
      required: true
      schema:
        type: string
      description: Agent identifier
...
  schemas:
    Agent:
      type: object
      required:
        - id
        - name
        - url
        - createdAt
      properties:
        id:
          type: string
          example: "agent_123"
        name:
          type: string
          example: "Financial Analysis Agent"
        url:
          type: string
          format: uri
          example: "https://api.dify.ai/v1"
        apiToken:
          type: string
          description: Encrypted API token
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        isActive:
          type: boolean
          default: true
    AgentCreate:
      type: object
      required:
        - name
        - url
        - apiToken
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        url:
          type: string
          format: uri
        apiToken:
          type: string
          minLength: 1
    AgentUpdate:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
          format: uri
        apiToken:
          type: string
        isActive:
          type: boolean
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "unauthorized"
            message: "Invalid or expired session"
    BadRequestError:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "bad_request"
            message: "Validation failed"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "not_found"
            message: "Resource does not exist"
...